requirements.txt
langchain
openai
python-dotenv
playwright